Lint and reorganize source code.
