--- AsyncReadFile/AsyncReadFile.txt
@@ -1,7 +1,0 @@
-dsdsds
-dsdsd
-dsd
-s
-dsdsd
-
-dsds
