+++ Timer/Timer1/Timer1.cpp
@@ -1,4 +1,3 @@
-#include <SDKDDKVer.h>
 #include <iostream>
 #include <boost/asio.hpp>
 #include <boost/asio/steady_timer.hpp>
