+++ Timer/Timer4/Timer4.cpp
@@ -1,4 +1,3 @@
-#include <SDKDDKVer.h>
 #include <iostream>
 #include <boost/asio.hpp>
 #include <boost/asio/steady_timer.hpp>
