+++ async-tcp-chatting/server/ServerMain.cpp
@@ -1,8 +1,4 @@
-#include <SDKDDKVer.h>
-
-
 #include "ChattingServer.h"
-
 
 
 const int MAX_SESSION_COUNT = 100;
