+++ async-tcp-echo/client/AsynchronousTCPClient.cpp
@@ -1,4 +1,4 @@
-#include <SDKDDKVer.h>
+
 #include <iostream>
 #include <boost/bind.hpp>
 #include <boost/asio.hpp>
