+++ resolver/resolver.cpp
@@ -1,5 +1,3 @@
-#include <SDKDDKVer.h>
-
 #include <iostream>
 #include <boost/asio.hpp>
 #include <boost/bind.hpp>
