+++ sync-tcp-echo/SynchronousTCPServer/SynchronousTCPServer.cpp
@@ -1,4 +1,3 @@
-#include <SDKDDKVer.h>
 #include <iostream>
 #include <boost/asio.hpp>
